API and Category

+++ src/main/java/com/Demo_Project_ECommerce/Demo_Project_E_Commerce/Repositories/AddressRepository/Address_Repository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface Address_Repository extends JpaRepository<Address,Integer> {
+public interface Address_Repository extends JpaRepository<Address,Long> {
 }
